Migrate Azure database controllers to managed resource reconciler
I was having a very difficult time trying to update the old gomega integration
style tests for the Azure databases and decided it would be easier to port them
to the new reconciler. This controller should be functionally equivalent to the
old one - i.e. it is not v1beta1 quality. It does not support updates, and has
no change to its schema, but should now be easier to update to v1beta1 quality.

Signed-off-by: Nic Cope <negz@rk0n.org>